requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
jinja2>=3.1.0
python-multipart>=0.0.9
pyyaml>=6.0
sqlalchemy>=2.0
passlib[bcrypt]>=1.7.4
PyJWT>=2.8.0
python-dotenv>=1.0
